refactor: split out SearchBarDropdown.tsx
